refactor(game): split 615-line help page into focused files under 250 lines
Reduces app/game/help/page.tsx from 615 lines to 87 by extracting:

- 8 presentation components into app/game/help/_components/
  - Section — shared section heading wrapper
  - TabBar — tab nav with aria-tablist semantics
  - tabs/{Overview,Phases,Castes,Combat,World,Contributor}Tab — one
    file per tab; pure static content

- 1 lib module into app/game/help/_lib/
  - tabs.ts — TABS list + TabId type + resolveTabId helper

The _components and _lib prefixes are Next.js private folders, so the new
structure does not introduce any new routes. page.tsx is now a thin
Suspense wrapper around the tab router that picks the right tab component
from the ?tab= query param.

No behavior change. All 1581 tests pass; tsc clean; lint clean.
